novas alterações

+++ horas-core-domain/src/main/java/br/com/wryel/horas/business/DemandaBusiness.java
@@ -6,6 +6,6 @@ import br.com.wryel.horas.entity.Demanda;
 import br.com.wryel.horas.entity.filter.DemandaFilter;
 
 @Local
-public interface DemandaBusiness extends Business<Demanda, Integer, DemandaFilter> {
+public interface DemandaBusiness extends Business<Demanda, Long, DemandaFilter> {
 	
 }
