+++ horas-core-domain/src/main/java/br/com/wryel/horas/dao/DemandaDAO.java
@@ -6,6 +6,6 @@ import br.com.wryel.horas.entity.Demanda;
 import br.com.wryel.horas.entity.filter.DemandaFilter;
 
 @Local
-public interface DemandaDAO extends DAO<Demanda, Integer, DemandaFilter> {
+public interface DemandaDAO extends DAO<Demanda, Long, DemandaFilter> {
 	
 }
